Use same insertion order logic
