--- packages/core/src/indexing/sqlite/__tests__/insertion-order.fixture.csv
@@ -1,22 +1,0 @@
-﻿stream_id;controller;last_anchored_at;created_at
-k2t6wysde758d3muw7okc3zspjlcpcjqzrchrdk5augruhpdha2od8tw46r0qh;did:key:foo;;9820
-k2t6wysde758d3tvke5zmqgq85s6m6dm593y5xl0desja6s1t6ub02osbsb8jv;did:key:foo;9900;9830
-k2t6wysde7589xsdd2tgrnedgm76fgk59cuo92pjibn6k9ichcxepmys19lluj;did:key:foo;9910;9840
-k2t6wysde758a7cnbum8auockii5f626vbi18keqegi52mob9z3kulecyvr04c;did:key:foo;9920;9850
-k2t6wysde758c8jmww7tcvcprvs6i7uxl30cz17jwssf4i1q5f96bmx2xkk65b;did:key:foo;9930;9860
-k2t6wysde758b3nriettrnxt5eg39ylwo2o7sbvie1ny5tprdmvwnx9xoqwy67;did:key:foo;9870;9870
-k2t6wysde758cgke6n0sr3yaor6shv5czq9d3dgmvb09jo8vy0gnlegbyowgs8;did:key:foo;9880;9880
-k2t6wyfsu4pfzxkvkqs4sxhgk2vy60icvko3jngl56qzmdewud4lscf5p93wna;did:key:foo;;9885
-k2t6wysde7589fze9445oe7y578f7csh4uexfqe9r9qziqyz8koegst94ryoat;did:key:foo;9860;9890
-k2t6wysde758a5g0poyqkp0ni2zfzw27j4fb2lkv5mkby35ss44kepa71axkk9;did:key:foo;9940;9900
-k2t6wysde758c9utjgo3caetaezu2cluj2p3gjttfqz111neyncd6z7ekxjfv9;did:key:foo;9990;9910
-k2t6wysde7589p6n2wl0yrh0mi0g6vyzrfd8xmo6wj8f5xbjjs3flxc8z7rzhm;did:key:foo;9950;9915
-k2t6wysde758cpnjmqwp3zcsrcthn116vq6ayeivfsivilr02pj6pybbuzgkbt;did:key:foo;10000;9920
-k2t6wysde7589wo4k3reakai9t5z52jgwfxdx8w0ug7px0s66o0h4qkkz134lb;did:key:foo;;9930
-k2t6wysde7589n4wvaiuhbokdvhtm71lxii24u8f12rlntsqakblz5u69vxgbu;did:key:blah;;9940
-k2t6wysde758et54lsbq54efgt73xqfg6s5sp72v9ervbn11z7w9rgb2r1fcu3;did:key:foo;;9950
-k2t6wysde758cx9qi61w2001a5l5hx8f5no9yue3ewv0ftky5mbvalmuf3gchb;did:key:foo;9970;9960
-k2t6wysde758aef79049queu07hgygp52a6i6an7of9y31vjf2bmyk2k0jwuod;did:key:foo;9960;9970
-k2t6wysde7589pfiw4q6rm96t2yzjiheopr56gx7jsqatip8qk9ebesdft1w14;did:key:foo;9980;9980
-k2t6wysde758akkonpg6flj8fitylax3fk40xrb9ud4hmivd29jdc097ad6nwz;did:key:foo;;9990
-k2t6wysde758e731xife7twg5dwpz8jg42vshm79ax6w7s0yu75kiaj8k2w6dt;did:key:blah;;10000
